# simd: add l2 byte AVX256 implementation