--- a/adapters/repos/db/vector/hnsw/distancer/asm/l2_amd64.go
+++ b/adapters/repos/db/vector/hnsw/distancer/asm/l2_amd64.go
@@ -43,6 +43,48 @@ func L2AVX512(x []float32, y []float32) float32 {
 }
 
 func L2ByteAVX256(x []uint8, y []uint8) uint32 {
+	switch len(x) {
+	case 1:
+		diff := uint32(x[0]) - uint32(y[0])
+		return diff * diff
+	case 2:
+		return l22[uint8, uint32](x, y)
+	case 3:
+		return l23[uint8, uint32](x, y)
+	case 4:
+		return l24[uint8, uint32](x, y)
+	case 5:
+		return l25[uint8, uint32](x, y)
+	case 6:
+		// manually inlined l26(x, y)
+		diff := uint32(x[5]) - uint32(y[5])
+		sum := diff * diff
+
+		diff = uint32(x[4]) - uint32(y[4])
+		sum += diff * diff
+
+		return l24[uint8, uint32](x, y) + sum
+	case 8:
+		// manually inlined l28(x, y)
+		diff := uint32(x[7]) - uint32(y[7])
+		sum := diff * diff
+
+		diff = uint32(x[6]) - uint32(y[6])
+		sum += diff * diff
+
+		diff = uint32(x[5]) - uint32(y[5])
+		sum += diff * diff
+
+		diff = uint32(x[4]) - uint32(y[4])
+		sum += diff * diff
+
+		return l24[uint8, uint32](x, y) + sum
+	case 10:
+		return l210[uint8, uint32](x, y)
+	case 12:
+		return l212[uint8, uint32](x, y)
+	}
+
 	var res uint32
 
 	l := len(x)