--- a/adapters/repos/db/vector/hnsw/distancer/asm/l2_inline_test.go
+++ b/adapters/repos/db/vector/hnsw/distancer/asm/l2_inline_test.go
@@ -53,32 +53,32 @@ func BenchmarkL2InlineVsLoop(b *testing.B) {
 				case 2:
 					b.ResetTimer()
 					for i := 0; i < b.N; i++ {
-						l22[float32, float32](x, y)
+						l22(x, y)
 					}
 				case 4:
 					b.ResetTimer()
 					for i := 0; i < b.N; i++ {
-						l24[float32, float32](x, y)
+						l24(x, y)
 					}
 				case 6:
 					b.ResetTimer()
 					for i := 0; i < b.N; i++ {
-						l26[float32, float32](x, y)
+						l26(x, y)
 					}
 				case 8:
 					b.ResetTimer()
 					for i := 0; i < b.N; i++ {
-						l28[float32, float32](x, y)
+						l28(x, y)
 					}
 				case 10:
 					b.ResetTimer()
 					for i := 0; i < b.N; i++ {
-						l210[float32, float32](x, y)
+						l210(x, y)
 					}
 				case 12:
 					b.ResetTimer()
 					for i := 0; i < b.N; i++ {
-						l212[float32, float32](x, y)
+						l212(x, y)
 					}
 				default:
 					panic("unsupported length")