--- a/adapters/repos/db/vector/hnsw/distancer/c/l2_byte_avx256_amd64.c
+++ b/adapters/repos/db/vector/hnsw/distancer/c/l2_byte_avx256_amd64.c
@@ -12,9 +12,6 @@
 #include <immintrin.h>
 #include <stdint.h>
 
-#include <immintrin.h>
-#include <stdint.h>
-
 void l2_byte_256(unsigned char *a, unsigned char *b, unsigned int *res, long *len)
 {
     int n = *len;