--- a/adapters/repos/db/vector/hnsw/distancer/l2_amd64_test.go
+++ b/adapters/repos/db/vector/hnsw/distancer/l2_amd64_test.go
@@ -17,6 +17,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer/asm"
 	"golang.org/x/sys/cpu"
 )
@@ -27,6 +28,17 @@ func L2PureGo(a, b []float32) float32 {
 	for i := range a {
 		diff := a[i] - b[i]
 		sum += diff * diff
+	}
+
+	return sum
+}
+
+func L2BytePureGo(a, b []uint8) uint32 {
+	var sum uint32
+
+	for i := range a {
+		diff := int32(a[i]) - int32(b[i])
+		sum += uint32(diff * diff)
 	}
 
 	return sum
@@ -78,6 +90,46 @@ func Test_L2_DistanceImplementation_OneNegativeValue(t *testing.T) {
 				asmResult = asm.L2AVX512(x, y)
 				assert.InEpsilon(t, control, asmResult, 0.01)
 			}
+		})
+	}
+}
+
+func Test_L2_Byte_DistanceImplementation(t *testing.T) {
+	lengths := []int{1, 4, 16, 31, 32, 35, 64, 67, 128, 130, 256, 260, 384, 390, 768, 777}
+
+	for _, length := range lengths {
+		t.Run(fmt.Sprintf("with vector l=%d", length), func(t *testing.T) {
+			x := make([]uint8, length)
+			y := make([]uint8, length)
+			for i := range x {
+				x[i] = uint8(rand.Uint32() % 256)
+				y[i] = uint8(rand.Uint32() % 256)
+			}
+
+			control := L2BytePureGo(x, y)
+
+			asmResult := asm.L2ByteAVX256(x, y)
+			require.Equal(t, int(control), int(asmResult))
+		})
+	}
+}
+
+func Test_L2_Byte_DistanceImplementation_OneNegativeValue(t *testing.T) {
+	lengths := []int{1, 4, 16, 31, 32, 35, 64, 67, 128, 130, 256, 260, 384, 390, 768, 777}
+
+	for _, length := range lengths {
+		t.Run(fmt.Sprintf("with vector l=%d", length), func(t *testing.T) {
+			x := make([]uint8, length)
+			y := make([]uint8, length)
+			for i := range x {
+				x[i] = uint8(rand.Uint32() % 256)
+				y[i] = uint8(rand.Uint32() % 256)
+			}
+
+			control := L2BytePureGo(x, y)
+
+			asmResult := asm.L2ByteAVX256(x, y)
+			require.Equal(t, int(control), int(asmResult))
 		})
 	}
 }
@@ -141,6 +193,68 @@ func Benchmark_L2(b *testing.B) {
 					asm.L2AVX512(x, y)
 				}
 			})
+		})
+	}
+}
+
+func Benchmark_L2Byte(b *testing.B) {
+	lengths := []int{
+		1,
+		2,
+		3,
+		4,
+		5,
+		6,
+		8,
+		10,
+		12,
+		16,
+		24,
+		30,
+		31,
+		32,
+		64,
+		67,
+		128,
+		256,
+		260,
+		299,
+		300,
+		384,
+		390,
+		600,
+		768,
+		777,
+		784,
+		1024,
+		1536,
+	}
+	for _, length := range lengths {
+		b.Run(fmt.Sprintf("vector dim=%d", length), func(b *testing.B) {
+			x := make([]uint8, length)
+			y := make([]uint8, length)
+			for i := range x {
+				x[i] = uint8(rand.Uint32() % 256)
+				y[i] = uint8(rand.Uint32() % 256)
+			}
+
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				// L2BytePureGo(x, y)
+				asm.L2ByteAVX256(x, y)
+			}
+
+			// b.Run("pure go", func(b *testing.B) {
+			// 	for i := 0; i < b.N; i++ {
+			// 		L2BytePureGo(x, y)
+			// 	}
+			// })
+
+			// b.Run("asm AVX", func(b *testing.B) {
+			// 	for i := 0; i < b.N; i++ {
+			// 		asm.L2ByteAVX256(x, y)
+			// 	}
+			// })
 		})
 	}
 }