… adapters/repos/db/vector/hnsw/distancer/asm/l2_byte_avx256_amd64.go
… adapters/repos/db/vector/hnsw/distancer/asm/l2_byte_avx256_amd64.s