### What's being changed:

```
goos: linux
goarch: amd64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer
cpu: Intel(R) Core(TM) i7-8700K CPU @ 3.70GHz
                          │ 1_l2_byte_go.txt │           2_l2_avx256.txt            │
                          │      sec/op      │    sec/op     vs base                │
_L2Byte/vector_dim=1-4          0.9604n ± 4%   3.2065n ± 3%  +233.89% (p=0.002 n=6)
_L2Byte/vector_dim=2-4           1.729n ± 2%    3.634n ± 3%  +110.18% (p=0.002 n=6)
_L2Byte/vector_dim=3-4           2.368n ± 1%    4.080n ± 1%   +72.35% (p=0.002 n=6)
_L2Byte/vector_dim=4-4           2.781n ± 2%    4.296n ± 2%   +54.49% (p=0.002 n=6)
_L2Byte/vector_dim=5-4           3.269n ± 1%    4.714n ± 2%   +44.20% (p=0.002 n=6)
_L2Byte/vector_dim=6-4           3.626n ± 1%    5.183n ± 1%   +42.96% (p=0.002 n=6)
_L2Byte/vector_dim=8-4           4.587n ± 2%    5.828n ± 2%   +27.05% (p=0.002 n=6)
_L2Byte/vector_dim=10-4          5.394n ± 1%    6.724n ± 3%   +24.66% (p=0.002 n=6)
_L2Byte/vector_dim=12-4          6.225n ± 2%    7.450n ± 3%   +19.68% (p=0.002 n=6)
_L2Byte/vector_dim=16-4          7.950n ± 1%    5.182n ± 1%   -34.83% (p=0.002 n=6)
_L2Byte/vector_dim=24-4         11.340n ± 2%    9.045n ± 1%   -20.23% (p=0.002 n=6)
_L2Byte/vector_dim=30-4          13.94n ± 4%    11.70n ± 1%   -16.10% (p=0.002 n=6)
_L2Byte/vector_dim=31-4          14.36n ± 1%    11.93n ± 2%   -16.92% (p=0.002 n=6)
_L2Byte/vector_dim=32-4         14.805n ± 2%    4.073n ± 3%   -72.49% (p=0.002 n=6)
_L2Byte/vector_dim=64-4         28.435n ± 1%    5.020n ± 2%   -82.35% (p=0.002 n=6)
_L2Byte/vector_dim=67-4         29.780n ± 4%    6.738n ± 3%   -77.37% (p=0.002 n=6)
_L2Byte/vector_dim=128-4        59.355n ± 2%    6.906n ± 3%   -88.36% (p=0.002 n=6)
_L2Byte/vector_dim=256-4        115.60n ± 2%    14.19n ± 1%   -87.72% (p=0.002 n=6)
_L2Byte/vector_dim=260-4        117.10n ± 1%    12.91n ± 2%   -88.97% (p=0.002 n=6)
_L2Byte/vector_dim=299-4        133.20n ± 1%    16.82n ± 3%   -87.37% (p=0.002 n=6)
_L2Byte/vector_dim=300-4        133.90n ± 1%    17.32n ± 1%   -87.06% (p=0.002 n=6)
_L2Byte/vector_dim=384-4        170.05n ± 3%    14.80n ± 1%   -91.30% (p=0.002 n=6)
_L2Byte/vector_dim=390-4        172.85n ± 1%    17.21n ± 4%   -90.04% (p=0.002 n=6)
_L2Byte/vector_dim=600-4        263.40n ± 1%    31.32n ± 2%   -88.11% (p=0.002 n=6)
_L2Byte/vector_dim=768-4        334.35n ± 1%    25.67n ± 2%   -92.32% (p=0.002 n=6)
_L2Byte/vector_dim=777-4        337.80n ± 2%    29.93n ± 1%   -91.14% (p=0.002 n=6)
_L2Byte/vector_dim=784-4        340.85n ± 3%    32.42n ± 1%   -90.49% (p=0.002 n=6)
_L2Byte/vector_dim=1024-4       442.60n ± 1%    33.06n ± 1%   -92.53% (p=0.002 n=6)
_L2Byte/vector_dim=1536-4       662.30n ± 2%    50.63n ± 1%   -92.35% (p=0.002 n=6)
geomean                          29.80n         10.17n        -65.89%
```

### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [ ] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.